requirements.txt
requests
python-telegram-bot==13.15
pandas
schedule
python-dotenv
SQLAlchemy
